requirements.txt
numpy==1.24.4
pytest==7.4.0
pycryptodome==3.18.0